refactor(fase-6): extract CloseRegister sub-components, shared NoRegisterModal, a11y fixes
- Extract ReconciliationSummaryCard, AmountInputCard, ApprovalModal from CloseRegister.jsx (702→~230 lines)
- Add shared NoRegisterModal component used by Sales and CloseRegister
- Fix: onKeyPress (deprecated) → onKeyDown in CancelTransactions
- Fix: remove console.log in CancelTransactions and OpenRegister
- Fix: add aria-label to action buttons, aria-hidden to decorative FA icons
- Fix: add role=alert/status to Alert components
- Fix: replace ⚠ emoji with FontAwesomeIcon faExclamationTriangle
- Fix: remove unused HiExclamationCircle import in ApprovalModal
- Lint: 0 errors, 0 warnings
