+++ src/pages/dashboard/Sales.jsx
@@ -2,11 +2,9 @@ import { useState, useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { transactionSchema, BONO_PAYMENT_METHOD_ID } from '../../utils/transactionSchema';
-import { Button, Card, Alert, Modal } from 'flowbite-react';
+import { Button, Card, Alert } from 'flowbite-react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faFileInvoice, faTimes } from '@fortawesome/free-solid-svg-icons';
-import { HiExclamationCircle } from 'react-icons/hi';
-import { useNavigate } from 'react-router-dom';
 import { getPaymentMethods, getPrevisions, getProfessionals, createTransaction } from '../../api';
 import { useStore } from '../../app/store';
 import SalesHeader from '../../components/Sales/SalesHeader';
@@ -17,9 +15,9 @@ import NotesSection from '../../components/Sales/NotesSection';
 import SummaryCard from '../../components/Sales/SummaryCard';
 import ConfirmModal from '../../components/Sales/ConfirmModal';
 import ReceiptModal from '../../components/Sales/ReceiptModal';
+import NoRegisterModal from '../../components/Commons/NoRegisterModal';
 
 const Sales = () => {
-  const navigate = useNavigate();
   const [showFolioInput, setShowFolioInput] = useState(false);
   const [paymentMethods, setPaymentMethods] = useState([]);
   const [previsions, setPrevisions] = useState([]);
@@ -278,59 +276,7 @@ const Sales = () => {
           <ReceiptModal open={showReceipt} onClose={() => setShowReceipt(false)} data={receiptData} />
         </Card>
 
-        {/* Modal de Caja No Abierta */}
-        <Modal show={showNoRegisterModal} onClose={() => {}} dismissible={false} size="md">
-          <Modal.Header>
-            <div className="flex items-center gap-2">
-              <HiExclamationCircle className="size-6 text-yellow-500" />
-              <span>Caja No Abierta</span>
-            </div>
-          </Modal.Header>
-          <Modal.Body>
-            <div className="space-y-4">
-              <Alert color="warning" icon={HiExclamationCircle}>
-                <span className="font-medium">Atención:</span> No hay una caja abierta actualmente.
-              </Alert>
-              
-              <div className="rounded-lg border border-gray-200 bg-gray-50 p-4 dark:border-slate-600 dark:bg-slate-700">
-                <p className="text-sm text-gray-700 dark:text-gray-300">
-                  Para realizar ventas y registrar transacciones, primero debe abrir una caja registradora. 
-                  La apertura de caja le permite:
-                </p>
-                <ul className="ml-4 mt-2 list-disc space-y-1 text-sm text-gray-600 dark:text-gray-400">
-                  <li>Registrar el monto inicial del día</li>
-                  <li>Asociar todas las ventas a la caja</li>
-                  <li>Mantener control de ingresos y egresos</li>
-                  <li>Realizar el cierre diario correctamente</li>
-                </ul>
-              </div>
-
-              <div className="dark:bg-secondary-900/20 rounded-lg bg-secondary-50 p-4">
-                <p className="text-sm text-secondary-800 dark:text-secondary-200">
-                  <span className="font-medium">¿Desea abrir una caja ahora?</span>
-                  <br />
-                  Será redirigido a la página de apertura de caja para comenzar.
-                </p>
-              </div>
-            </div>
-          </Modal.Body>
-          <Modal.Footer>
-            <div className="flex w-full justify-end gap-3">
-              <Button
-                color="gray"
-                onClick={() => navigate('/dashboard')}
-              >
-                Cancelar
-              </Button>
-              <Button
-                color="success"
-                onClick={() => navigate('/dashboard/open-register')}
-              >
-                Abrir Caja
-              </Button>
-            </div>
-          </Modal.Footer>
-        </Modal>
+        <NoRegisterModal show={showNoRegisterModal} context="sales" />
       </div>
     </div>
   );
